+++ src/assembly_language_provider.ts
@@ -1,6 +1,11 @@
 import * as vscode from 'vscode';
 import * as childProcess from 'child_process';
 import * as path from 'path';
+import * as fs from 'fs';
+import * as os from 'os';
+import { Disassemble } from './backend/profile';
+import { GetCycles } from './client/68k';
+import { GetCpuDoc, GetCpuInsns, GetCpuName } from './client/docs';
 
 enum TokenTypes {
 	function,
@@ -19,6 +24,14 @@ function split_with_offset(str: string, re: RegExp) {
 	return results;
 }
 
+export function getEditorForDocument(doc: vscode.TextDocument): vscode.TextEditor {
+	for(const textEditor of vscode.window.visibleTextEditors) {
+		if(textEditor.document === doc)
+			return textEditor;
+	}
+	return null;
+}
+
 interface Token {
 	line: number;
 	char: number;
@@ -29,6 +42,7 @@ interface Token {
 class SourceContext {
 	public static extensionPath: string;
 	public labels = new Map<string, number>(); // label -> line
+	public cycles = new Map<number, string>(); // line -> cycles text
 	public tokens: Token[] = [];
 	private text: string;
 
@@ -40,12 +54,21 @@ class SourceContext {
 	}
 
 	public parse() {
-		const objdump = childProcess.spawnSync(
+		const date = new Date();
+		const dateString = date.getFullYear().toString() + "." + (date.getMonth()+1).toString().padStart(2, '0') + "." + date.getDate().toString().padStart(2, '0') + "-" +
+			date.getHours().toString().padStart(2, '0') + "." + date.getMinutes().toString().padStart(2, '0') + "." + date.getSeconds().toString().padStart(2, '0');
+		
+		const tmp = path.join(os.tmpdir(), `amiga-as-${dateString}.o`);
+		try {
+			fs.unlinkSync(tmp);
+		} catch(e) {}
+		const as = childProcess.spawnSync(
 			path.join(SourceContext.extensionPath, "bin/opt/bin/m68k-amiga-elf-as.exe"), 
 			[
 				'-', // input from stdin
-				'-o', 'nul', // no object output
+				'-o', tmp, // no object output
 				'--register-prefix-optional', 
+				'-g', // debug info
 				'-asn', // enable listing to stdout; 's' = symbol table, 'n' = turn off forms
 				'-L' // include local labels
 			], 
@@ -54,8 +77,8 @@ class SourceContext {
 				maxBuffer: 10*1024*1024 
 			});
 
-		const stdout = objdump.stdout.toString().replace(/\r/g, '').split('\n');
-		const stderr = objdump.stderr.toString().replace(/\r/g, '').split('\n');
+		const stdout = as.stdout.toString().replace(/\r/g, '').split('\n');
+		const stderr = as.stderr.toString().replace(/\r/g, '').split('\n');
 
 		// get labels
 		this.labels.clear();
@@ -77,7 +100,85 @@ class SourceContext {
 		}
 		this.diagnosticCollection.set(vscode.Uri.file(this.fileName), errors);
 
+		// get cycles from disassembly
+		this.cycles.clear();
+		if(fs.existsSync(tmp)) {
+			const objdumpPath = path.join(SourceContext.extensionPath, "bin/opt/bin/m68k-amiga-elf-objdump.exe");
+			try {
+				const objdump = Disassemble(objdumpPath, tmp);
+				// from objdump.tsx
+				const lines = objdump.replace(/\r/g, '').split('\n');
+				let lineNum = 0;
+				for(const line of lines) {
+					const locMatch = line.match(/^(\S.+):([0-9]+)( \(discriminator [0-9]+\))?$/); // C:/Users/Chuck/Documents/Visual_Studio_Code/amiga-debug/template/support/gcc8_c_support.c:62 (discriminator 1)
+					if(locMatch) {
+						lineNum = parseInt(locMatch[2]);
+						continue;
+					}
+					//                                PC             HEX WORDS           OPCODE REST
+					const insnMatch = line.match(/^ *([0-9a-f]+):\t((?:[0-9a-f]{4} )+)\s*(\S+)(?:\s(.*))?$/); //      cce:	0c40 a00e      	cmpi.w #-24562,d0
+					if(insnMatch) {
+						const pc = parseInt(insnMatch[1], 16);
+						const hex = insnMatch[2].split(' ');
+						const opcode = insnMatch[3];
+						const rest = insnMatch[4] || '';
+						const insn = new Uint16Array(hex.length);
+						hex.forEach((h, i) => { insn[i] = parseInt(h, 16); });
+						if(lineNum !== 0)
+							this.cycles.set(lineNum, GetCycles(insn).map((c) => `${c.total}`).join('-') + 'T');
+						lineNum = 0;
+					}
+				}
+			} catch(e) {}
+			try {
+				fs.unlinkSync(tmp);
+			} catch(e) {}
+		}
+
 		this.tokens = this.getTokens();
+	}
+
+	// theoretical 68000 cycle decorations
+	private static decoration = vscode.window.createTextEditorDecorationType({
+		before: {
+			textDecoration: 'none; white-space: pre; border-radius: 6px; padding: 0 10px 0 10px',
+			backgroundColor: new vscode.ThemeColor("badge.background"),
+			color: new vscode.ThemeColor("badge.foreground"),
+			margin: '0 10px 0 10px'
+		}
+	});
+	private static decorationEmpty = vscode.window.createTextEditorDecorationType({
+		before: {
+			textDecoration: 'none; white-space: pre; padding: 0 10px 0 10px',
+			margin: '0 10px 0 10px',
+			contentText: '      '
+		}
+	});
+
+	public setDecorations(textEditor: vscode.TextEditor) {
+		if(textEditor === null)
+			return;
+		
+		const optionsArray: vscode.DecorationOptions[] = [];
+		const emptyRanges: vscode.Range[] = [];
+		for(let line = 1; line < textEditor.document.lineCount; line++) {
+			const cyclesStr = this.cycles.get(line);
+			const range = new vscode.Range(new vscode.Position(line - 1, 0), new vscode.Position(line - 1, 0));
+			if(cyclesStr !== undefined) {
+				optionsArray.push({ 
+					range, 
+					renderOptions: { 
+						before: { 
+							contentText: cyclesStr.padStart(6, ' ')
+						} 
+					} 
+				});
+			} else {
+				emptyRanges.push(range);
+			}
+		}
+		textEditor.setDecorations(SourceContext.decoration, optionsArray);
+		textEditor.setDecorations(SourceContext.decorationEmpty, emptyRanges);
 	}
 
 	private getTokens(): Token[] {
@@ -97,7 +198,7 @@ class SourceContext {
 	}
 }
 
-export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvider, vscode.DefinitionProvider, vscode.DocumentSymbolProvider {
+export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvider, vscode.DefinitionProvider, vscode.DocumentSemanticTokensProvider, vscode.CompletionItemProvider, vscode.HoverProvider {
 	private sourceContexts = new Map<string, SourceContext>();
 	public diagnosticCollection: vscode.DiagnosticCollection = vscode.languages.createDiagnosticCollection(AmigaAssemblyLanguageProvider.getLanguageId());
 
@@ -108,7 +209,7 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 	constructor(extensionPath: string) {
 		SourceContext.extensionPath = extensionPath;
 
-		const changeTimers = new Map<string, any>(); // Keyed by file name.
+		const changeTimers = new Map<string, NodeJS.Timeout>(); // Keyed by file name.
 
 		// parse documents that are already open when extension is activated
 		for(const document of vscode.workspace.textDocuments) {
@@ -116,6 +217,7 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 				console.log("initial parse " + document.fileName);
 				this.getSourceContext(document.fileName).setText(document.getText());
 				this.getSourceContext(document.fileName).parse();
+				this.getSourceContext(document.fileName).setDecorations(getEditorForDocument(document));
 			}
 		}
 
@@ -125,6 +227,7 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 				console.log("openTextDocument: initial parse " + document.fileName);
 				this.getSourceContext(document.fileName).setText(document.getText());
 				this.getSourceContext(document.fileName).parse();
+				this.getSourceContext(document.fileName).setDecorations(getEditorForDocument(document));
 			}
 		});
 
@@ -140,12 +243,13 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 					changeTimers.delete(fileName);
 					console.log("reparse " + event.document.fileName);
 					this.getSourceContext(event.document.fileName).parse();
+					this.getSourceContext(event.document.fileName).setDecorations(getEditorForDocument(event.document));
 				}, 300));
 			}
 		});
 	}
 
-	private getSourceContext(fileName: string): SourceContext {
+	public getSourceContext(fileName: string): SourceContext {
 		let context = this.sourceContexts.get(fileName);
 		if(context)
 			return context;
@@ -154,6 +258,35 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 		return context;
 	}
 
+	// CompletionItemProvider
+	public provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext): vscode.ProviderResult<vscode.CompletionItem[] | vscode.CompletionList<vscode.CompletionItem>> {
+		const items: vscode.CompletionItem[] = [];
+		for(const opc of GetCpuInsns()) {
+			items.push(new vscode.CompletionItem({ label: opc, description: GetCpuName(opc) }, vscode.CompletionItemKind.Keyword));
+		}
+		const sourceContext = this.getSourceContext(document.fileName);
+		sourceContext.labels.forEach((value, key) => {
+			items.push(new vscode.CompletionItem({ label: key }, vscode.CompletionItemKind.Function));
+		});
+		return items;
+	}
+
+	public resolveCompletionItem?(item: vscode.CompletionItem, token: vscode.CancellationToken): vscode.ProviderResult<vscode.CompletionItem> {
+		return null;
+		//throw new Error('Method not implemented.');
+	}
+
+	// HoverProvider
+	public provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Hover> {
+		const word = document.getText(document.getWordRangeAtPosition(position));
+		console.log(word);
+		const doc = GetCpuDoc(word);
+		if(doc)
+			return new vscode.Hover(new vscode.MarkdownString(doc));
+		return null;
+	}
+
+	// DocumentSymbolProvider
 	public provideDocumentSymbols(document: vscode.TextDocument, token: vscode.CancellationToken): vscode.ProviderResult<vscode.SymbolInformation[] | vscode.DocumentSymbol[]> {
 		const context = this.getSourceContext(document.fileName);
 		const symbols: vscode.DocumentSymbol[] = [];
@@ -163,6 +296,7 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 		return symbols;
 	}
 
+	// DefinitionProvider	
 	public provideDefinition(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Location | vscode.Location[] | vscode.LocationLink[]> {
 		const context = this.getSourceContext(document.fileName);
 		const word = document.getText(document.getWordRangeAtPosition(position));
@@ -173,9 +307,10 @@ export class AmigaAssemblyLanguageProvider implements vscode.DocumentSymbolProvi
 		return new vscode.Location(document.uri, new vscode.Position(line, 0));
 	}
 
+	// DocumentSemanticTokensProvider
 	public static getSemanticTokensLegend() {
 		const tokenTypes = ['function'];
-		const tokenModifiers = [];
+		const tokenModifiers: string[] = [];
 		return new vscode.SemanticTokensLegend(tokenTypes, tokenModifiers);
 	}