+++ src/client/objdump.tsx
@@ -1,5 +1,6 @@
 import { Component, Fragment, FunctionComponent, h, JSX } from 'preact';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'preact/hooks';
+import { createPortal } from 'preact/compat';
 import { Cycles, GetCycles, GetJump, JumpType } from "./68k";
 import { DropdownComponent, DropdownOptionProps } from './dropdown';
 import { Icon } from './icons';
@@ -17,6 +18,8 @@ import { Absolute, VirtualList } from './virtual_list';
 import { VsCodeApi } from "./vscodeApi";
 import Highlighter from 'react-highlight-words';
 import { resolvePath } from './pathResolve';
+import { GetCpuDoc, GetCpuName, NormalizeInsn } from './docs';
+import Markdown from 'markdown-to-jsx';
 
 // messages from webview to vs code
 export interface IOpenDocumentMessageObjview {
@@ -43,6 +46,8 @@ export interface Line {
 	traceHits?: number;
 	traceCycles?: number;
 	theoreticalCycles?: Cycles[];
+	opcode?: string;
+	rest?: string;
 	text: string;
 	loc?: Location;
 }
@@ -59,7 +64,7 @@ export class ObjdumpModel {
 	public jumps: JumpInfo[] = [];
 
 	private addJumps(func: Function, funcJumps: JumpInfo[]) {
-		// based on binutils-gdb/binutils/objdump.c
+		//console.time('addJumps');
 		const sortedJumps = funcJumps.filter((j) => j.end >= func.pc && j.end < func.end).sort((a: JumpInfo, b: JumpInfo) => {
 			const aMin  = Math.min(...a.start, a.end);
 			const bMin  = Math.min(...b.start, b.end);
@@ -79,44 +84,25 @@ export class ObjdumpModel {
 			return aMax >= bMin && aMin <= bMax;
 		}
 
-		let lastJump = 0;
+		//console.log(`func: ${func.name}`);
 		let maxLevel = -1;
-		while(lastJump < sortedJumps.length) {
-			// The last jump is part of the next group
-			const base = lastJump;
-			// Increment level
-			sortedJumps[base].level = ++maxLevel;
-			// Find jumps that can be combined on the same level, with the largest jumps tested first.
-			// This has the advantage that large jumps are on lower levels and do not intersect with small
-			// jumps that get grouped on higher levels.
-			let exchangeItem = lastJump + 1;
-			for(let it = exchangeItem; it < sortedJumps.length; it++) {
-				// test if the jump intersects with any jump from current group
-				let ok = true;
-				for(let itCollision = base; itCollision !== exchangeItem; itCollision++) {
-					// this jump intersects so we leave it out
-					if(jumpIntersects(sortedJumps[itCollision], sortedJumps[it])) {
-						ok = false;
-						break;
-					}
-				}
-				// add jump to group
-				if(ok) {
-					// move current element to the front
-					if(it !== exchangeItem) {
-						// hmm.. this code is not good
-						[sortedJumps[exchangeItem], sortedJumps[it]] = [sortedJumps[it], sortedJumps[exchangeItem]];
-						lastJump = it;
-					} else {
-						lastJump = exchangeItem;
-						exchangeItem++;
-					}
-					sortedJumps[lastJump].level = maxLevel;
+		for(let i = 0; i < sortedJumps.length; i++) {
+			if(sortedJumps[i].level !== -1)
+				continue;
+			sortedJumps[i].level = ++maxLevel;
+			const currentGroup: number[] = [i];
+			for(let j = i + 1; j < sortedJumps.length; j++) {
+				if(sortedJumps[j].level !== -1)
+					continue;
+				if(currentGroup.every((e) => sortedJumps[j].level !== -1 || !jumpIntersects(sortedJumps[e], sortedJumps[j]))) {
+					sortedJumps[j].level = maxLevel;
+					currentGroup.push(j);
 				}
 			}
-			lastJump = exchangeItem; // move to next group
+			//console.log(`  level ${maxLevel}: ${JSON.stringify(currentGroup)}`);
 		}
 		this.jumps.push(...sortedJumps);
+		//console.timeEnd('addJumps');
 	}
 
 	constructor(objdump: string, theoreticalCycles = true, pcTrace: number[] = []) {
@@ -172,7 +158,9 @@ export class ObjdumpModel {
 				}
 				this.content.push({
 					pc,
-					text: `${pc.toString(16).padStart(8, ' ')}: ${opcode.padEnd(7, ' ')} ${rest}`, // ${insnMatch[2].padEnd(20, ' ')} 
+					text: `${pc.toString(16).padStart(8, ' ')}: ${opcode.padEnd(7, ' ')} ${rest}`,
+					opcode,
+					rest,
 					theoreticalCycles: theoreticalCycles ? GetCycles(insn) : undefined,
 					loc,
 					traceHits: hits.get(pc),
@@ -230,6 +218,7 @@ export const ObjdumpView: FunctionComponent<{
 				pcMap.set(line.pc, index);
 		});
 
+
 		const jumps: JumpAbsolute[] = model.jumps.filter((jump) =>
 			[...jump.start, jump.end].every((pc) => pcMap.has(pc))
 		).map((jump) => {
@@ -244,8 +233,6 @@ export const ObjdumpView: FunctionComponent<{
 				height: (max - min + 1) * rowHeight
 			};
 		});
-		//console.log(jumps);
-
 		return [model.content, model.functions.sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase())), jumps];
 	}, [model, rowHeight]);
 
@@ -319,6 +306,33 @@ export const ObjdumpView: FunctionComponent<{
 		});
 	}, []);
 
+	const [hovered, setHovered] = useState<{ markdown: string; x: number; y: number; justify: string}>({ markdown: '', x: -1, y: -1, justify: '' });
+	const tooltipRef = useRef<HTMLDivElement>();
+
+	const onMouseEnterOpcode = useCallback((evt: JSX.TargetedMouseEvent<HTMLSpanElement>) => {
+		// eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-member-access
+		const opcode = evt.currentTarget.attributes['data'].nodeValue as string;
+		const rect = evt.currentTarget.getBoundingClientRect();
+		const markdown = GetCpuDoc(opcode);
+		if(markdown) {
+			const hov = { 
+				markdown, 
+				x: Math.min(rect.left, window.innerWidth - 530), 
+				y: rect.bottom < window.innerHeight - 260 ? rect.bottom + 10 : rect.top - 260,
+				justify: rect.bottom < window.innerHeight - 260 ? 'flex-start' : 'flex-end' 
+			};
+			setHovered(hov);
+		}
+	}, []);
+	const onMouseLeaveOpcode = useCallback(() => {
+		setHovered({ markdown: '', x: -1, y: -1, justify: '' });
+	}, []);
+	const onWheelOpcode = useCallback((evt: WheelEvent) => {
+		evt.preventDefault();
+		// dunno how to make smooth scrolling that works when wheeling repeatedly
+		tooltipRef.current.scrollTop += evt.deltaY;
+	}, [tooltipRef.current]);
+
 	const renderRow = useCallback((c: Line, index: number) => {
 		const extra: string[] = [];
 
@@ -343,7 +357,13 @@ export const ObjdumpView: FunctionComponent<{
 
 		const text = (find.internal && findResult.length > 0) 
 		? <Highlighter searchWords={[find.text]} autoEscape={true} highlightClassName={styles.find_hit} textToHighlight={c.text} />
-		: c.text;
+		: (c.opcode 
+			? <Fragment>
+				<span>{c.pc.toString(16).padStart(8, ' ')}: </span>
+				<span class={styles.opcode} data={c.opcode} onMouseEnter={onMouseEnterOpcode} onMouseLeave={onMouseLeaveOpcode} onWheel={onWheelOpcode}>{c.opcode}</span>
+				<span>{' '.repeat(7 - c.opcode.length)} {c.rest}</span>
+			</Fragment> 
+			: c.text);
 
 		return (c.pc === undefined
 		? <div class={[styles.row, ...extra].join(' ')} data-row={index}>{text}{'\n'}</div>
@@ -358,7 +378,7 @@ export const ObjdumpView: FunctionComponent<{
 			{(c.loc !== undefined && frame !== -1) ? <div class={styles.file}><a href='#' data-file={c.loc.file} data-line={c.loc.line} onClick={onClickLoc}>{c.loc.file}:{c.loc.line}</a></div> : ''}
 			{'\n'}
 		</div>);
-	}, [onClickLoc, row, content, frame, findResult, find, curFind]);
+	}, [onClickLoc, row, content, frame, findResult, find, curFind, onMouseEnterOpcode, onMouseLeaveOpcode, onWheelOpcode]);
 
 	const renderJump = useCallback((jump: JumpAbsolute) => {
 		const right = 70; // needs to match CSS
@@ -381,7 +401,14 @@ export const ObjdumpView: FunctionComponent<{
 					const text = minCycles === maxCycles ? `${minCycles}T` : `${minCycles}-${maxCycles}T`;
 					return <text transform={`translate(${indent + 2}, ${endY + 3 + ((jump.start[0] - min) * rowHeight + rowMiddle - (endY + 3)) / 2}) rotate(-90)`} textAnchor="middle" dominant-baseline="hanging" class={styles.jumpduration} stroke="none">{text}</text>;
 				}
-			}
+			}/* else {
+				// debug
+				return <Fragment>
+					<text transform={`translate(${Math.max(15, indent + 2)}, ${((0) * rowHeight + rowMiddle)})`} textAnchor="middle" dominant-baseline="hanging" class={styles.jumpduration} stroke="none">{jump.level}</text>
+					<text transform={`translate(${indent + 2}, ${endY + 3 + ((jump.start[0] - min) * rowHeight + rowMiddle - (endY + 3)) / 2})`} textAnchor="middle" dominant-baseline="hanging" class={styles.jumpduration} stroke="none">{jump.level}</text>
+					<text transform={`translate(${Math.max(15, indent + 2)}, ${end * rowHeight})`} textAnchor="middle" dominant-baseline="hanging" class={styles.jumpduration} stroke="none">{jump.level}</text>
+				</Fragment>;
+			}*/
 			return '';
 		})();
 
@@ -621,5 +648,11 @@ export const ObjdumpView: FunctionComponent<{
 			<FunctionDropdown alwaysChange={true} options={functions} value={func} onChange={onChangeFunction} />
 		</div>
 		<VirtualListLine ref={listRef} class={styles.container} style={{opacity}} rows={content} renderRow={renderRow} rowHeight={rowHeight} absolutes={jumps} renderAbsolute={renderJump} overscanCount={50} onclick={onClickContainer} />
+		{hovered.markdown !== '' && (createPortal(
+			<div class={styles.tooltip_parent} style={{justifyContent: hovered.justify, left: hovered.x, top: hovered.y }}>
+				<div ref={tooltipRef} class={styles.tooltip}>
+					<Markdown>{hovered.markdown}</Markdown>
+				</div>
+			</div>, document.body))}
 	</Fragment>;
 };